Added casts to reduce warnings (based on report by Brad Hards).

+++ lib/gnutls_buffers.c
@@ -138,19 +138,19 @@ _gnutls_record_buffer_put (content_type_t type,
     case GNUTLS_APPLICATION_DATA:
       buf = &session->internals.application_data_buffer;
       _gnutls_buffers_log ("BUF[REC]: Inserted %d bytes of Data(%d)\n",
-			   length, type);
+			   (int)length, (int)type);
       break;
 
     case GNUTLS_HANDSHAKE:
       buf = &session->internals.handshake_data_buffer;
       _gnutls_buffers_log ("BUF[HSK]: Inserted %d bytes of Data(%d)\n",
-			   length, type);
+			   (int)length, (int)type);
       break;
 
     case GNUTLS_INNER_APPLICATION:
       buf = &session->internals.ia_data_buffer;
       _gnutls_buffers_log ("BUF[IA]: Inserted %d bytes of Data(%d)\n",
-			   length, type);
+			   (int)length, (int)type);
       break;
 
     default:
@@ -221,20 +221,20 @@ _gnutls_record_buffer_get (content_type_t type,
     case GNUTLS_APPLICATION_DATA:
       _gnutls_buffer_get_data( &session->internals.application_data_buffer, data, &length);
       _gnutls_buffers_log ("BUFFER[REC][AD]: Read %d bytes of Data(%d)\n",
-			   length, type);
+			   (int)length, (int)type);
       break;
 
     case GNUTLS_HANDSHAKE:
       _gnutls_buffer_get_data( &session->internals.handshake_data_buffer, data, &length);
       _gnutls_buffers_log ("BUF[REC][HD]: Read %d bytes of Data(%d)\n",
-			   length, type);
+			   (int)length, (int)type);
       break;
 
     case GNUTLS_INNER_APPLICATION:
 
       _gnutls_buffer_get_data( &session->internals.ia_data_buffer, data, &length);
       _gnutls_buffers_log ("BUF[REC][IA]: Read %d bytes of Data(%d)\n",
-			   length, type);
+			   (int)length, (int)type);
       break;
 
     default:
@@ -307,7 +307,7 @@ _gnutls_read (gnutls_session_t session, void *iptr,
 	    : errno;
 
 	  _gnutls_read_log ("READ: %d returned from %p, errno=%d gerrno=%d\n",
-			    i, fd, errno, session->internals.errnum);
+			    (int)i, fd, errno, session->internals.errnum);
 
 	  if (err == EAGAIN || err == EINTR)
 	    {
@@ -315,7 +315,7 @@ _gnutls_read (gnutls_session_t session, void *iptr,
 		{
 
 		  _gnutls_read_log ("READ: returning %d bytes from %p\n",
-				    sizeOfPtr - left, fd);
+				    (int)(sizeOfPtr - left), fd);
 
 		  goto finish;
 		}
@@ -334,7 +334,7 @@ _gnutls_read (gnutls_session_t session, void *iptr,
       else
 	{
 
-	  _gnutls_read_log ("READ: Got %d bytes from %p\n", i, fd);
+	  _gnutls_read_log ("READ: Got %d bytes from %p\n", (int)i, fd);
 
 	  if (i == 0)
 	    break;		/* EOF */
@@ -353,7 +353,7 @@ finish:
 
 
       _gnutls_read_log ("READ: read %d bytes from %p\n",
-			(sizeOfPtr - left), fd);
+			(int)(sizeOfPtr - left), fd);
 
       for (x = 0; x < ((sizeOfPtr - left) / 16) + 1; x++)
 	{
@@ -546,8 +546,8 @@ _gnutls_io_read_buffered (gnutls_session_t session, opaque ** iptr,
     {
       _gnutls_read_log
 	("RB: Have %d bytes into buffer. Adding %d bytes.\n",
-	 session->internals.record_recv_buffer.length, ret);
-      _gnutls_read_log ("RB: Requested %d bytes\n", sizeOfPtr);
+	 (int)session->internals.record_recv_buffer.length, (int)ret);
+      _gnutls_read_log ("RB: Requested %d bytes\n", (int)sizeOfPtr);
       session->internals.record_recv_buffer.length += ret;
     }
 
@@ -569,10 +569,10 @@ _gnutls_io_read_buffered (gnutls_session_t session, opaque ** iptr,
 
       if (ret2 > 0)
 	{
-	  _gnutls_read_log ("RB-PEEK: Read %d bytes in PEEK MODE.\n", ret2);
+	  _gnutls_read_log ("RB-PEEK: Read %d bytes in PEEK MODE.\n", (int)ret2);
 	  _gnutls_read_log
 	    ("RB-PEEK: Have %d bytes into buffer. Adding %d bytes.\nRB: Requested %d bytes\n",
-	     session->internals.record_recv_buffer.length, ret2, sizeOfPtr);
+	     (int)session->internals.record_recv_buffer.length, (int)ret2, (int)sizeOfPtr);
 	  session->internals.have_peeked_data = 1;
 	  session->internals.record_recv_buffer.length += ret2;
 
@@ -663,10 +663,10 @@ _gnutls_io_write_buffered (gnutls_session_t session,
       n = bdata.size;
 
       _gnutls_write_log
-	("WRITE: Restoring old write. (%d bytes to send)\n", n);
+	("WRITE: Restoring old write. (%d bytes to send)\n", (int)n);
     }
 
-  _gnutls_write_log ("WRITE: Will write %d bytes to %p.\n", n, fd);
+  _gnutls_write_log ("WRITE: Will write %d bytes to %p.\n", (int)n, fd);
 
   i = 0;
   left = n;
@@ -723,7 +723,7 @@ _gnutls_io_write_buffered (gnutls_session_t session,
 
 	      _gnutls_write_log
 		("WRITE: Interrupted. Stored %d bytes to buffer. Already sent %d bytes.\n",
-		 left, n - left);
+		 (int)left, (int)(n - left));
 
 	      if (err == EAGAIN)
 		return GNUTLS_E_AGAIN;
@@ -746,7 +746,7 @@ _gnutls_io_write_buffered (gnutls_session_t session,
 
 	  _gnutls_write_log
 	    ("WRITE: wrote %d bytes to %p. Left %d bytes. Total %d bytes.\n",
-	     i, fd, left, n);
+	     (int)i, fd, (int)left, (int)n);
 	  for (x = 0; x < (unsigned) ((i) / 16) + 1; x++)
 	    {
 	      line[0] = 0;
@@ -795,8 +795,8 @@ _gnutls_io_write_flush (gnutls_session_t session)
     return 0;			/* done */
 
   ret = _gnutls_io_write_buffered (session, NULL, 0);
-  _gnutls_write_log ("WRITE FLUSH: %d [buffer: %d]\n", ret,
-		     session->internals.record_send_buffer.length);
+  _gnutls_write_log ("WRITE FLUSH: %d [buffer: %d]\n", (int)ret,
+		     (int)session->internals.record_send_buffer.length);
 
   return ret;
 }
@@ -816,7 +816,7 @@ _gnutls_handshake_io_write_flush (gnutls_session_t session)
       return ret;
     }
 
-  _gnutls_write_log ("HANDSHAKE_FLUSH: written[1] %d bytes\n", ret);
+  _gnutls_write_log ("HANDSHAKE_FLUSH: written[1] %d bytes\n", (int)ret);
 
   if (session->internals.handshake_send_buffer.length == 0)
     {
@@ -1068,7 +1068,7 @@ _gnutls_handshake_buffer_put (gnutls_session_t session, opaque * data,
       return GNUTLS_E_HANDSHAKE_TOO_LARGE;
     }
 
-  _gnutls_buffers_log ("BUF[HSK]: Inserted %d bytes of Data\n", length);
+  _gnutls_buffers_log ("BUF[HSK]: Inserted %d bytes of Data\n", (int)length);
 
   if (_gnutls_buffer_append (&session->internals.handshake_hash_buffer,
 			     data, length) < 0)
@@ -1099,7 +1099,7 @@ _gnutls_handshake_buffer_peek (gnutls_session_t session, opaque * data,
       length = session->internals.handshake_hash_buffer.length;
     }
 
-  _gnutls_buffers_log ("BUF[HSK]: Peeked %d bytes of Data\n", length);
+  _gnutls_buffers_log ("BUF[HSK]: Peeked %d bytes of Data\n", (int)length);
 
   memcpy (data, session->internals.handshake_hash_buffer.data, length);
   return length;
@@ -1115,7 +1115,7 @@ _gnutls_handshake_buffer_get_ptr (gnutls_session_t session,
   if (length != NULL)
     *length = session->internals.handshake_hash_buffer.length;
 
-  _gnutls_buffers_log ("BUF[HSK]: Peeked %d bytes of Data\n", *length);
+  _gnutls_buffers_log ("BUF[HSK]: Peeked %d bytes of Data\n", (int)*length);
 
   if (data_ptr != NULL)
     *data_ptr = session->internals.handshake_hash_buffer.data;
